requirements.txt
psycopg2-binary
Flask
Werkzeug
python-dotenv